Renamed file from intepreterActivation to envrionmentActivationService (#8579)
